+++ lib/index.ts
@@ -1,4 +1,5 @@
 export * from './cleanup.decorator';
 export * from './done-subject';
 export * from './rx-util';
+export * from './state-subject';
 export * from './worker-observable';