+++ lib/state-subject.ts
@@ -1,0 +1,9 @@
+import {BehaviorSubject} from 'rxjs';
+
+export class StateSubject<T> extends BehaviorSubject<T> {
+  next(value: T) {
+    if (this.value !== value) {
+      super.next(value);
+    }
+  }
+}